Fix logic of query builder button
