+++ src/components/MainListInfoForm/MainListInfoForm.module.css
@@ -3,7 +3,7 @@
 }
 
 .recordTypeField {
-    width: 130px;
+    width: 170px;
 }
 
 .radioLabels {
