+++ src/interfaces/entity.ts
@@ -1,6 +1,15 @@
+import { ReactNode } from 'react';
+
 export interface Value {
     label: string;
     name: string;
+}
+
+export type EntityTypeOption = {
+    label: string | ReactNode,
+    value: string,
+    selected: boolean,
+    disabled?: boolean
 }
 
 export interface EntityTypeColumn {
